[Fix #10731] Show tip for suggested extensions that are installed but not loaded in .rubocop.yml
