+++ changelog/change_show_tip_for_suggested_extensions_that.md
@@ -1,0 +1,1 @@
+* [#10731](https://github.com/rubocop/rubocop/issues/10731): Show tip for suggested extensions that are installed but not loaded in .rubocop.yml. ([@nobuyo][])
