+++ spec/rubocop/cli_spec.rb
@@ -1811,309 +1811,6 @@ RSpec.describe RuboCop::CLI, :isolated_environment do
     end
   end
 
-  describe 'extension suggestions' do
-    matcher :suggest_extensions do |*extensions|
-      supports_block_expectations
-      attr_accessor :suggested
-
-      def suggestion_regex
-        Regexp.new(<<~REGEXP, Regexp::MULTILINE).freeze
-          Tip: Based on detected gems, the following RuboCop extension libraries might be helpful:
-          (?<suggestions>.*)
-        REGEXP
-      end
-
-      def find_suggestions
-        actual.call
-        suggestions = (suggestion_regex.match($stdout.string) || {})[:suggestions]
-        self.suggested = suggestions ? suggestions.scan(/(?<=\* )[a-z0-9_-]+\b/.freeze) : []
-      end
-
-      match do
-        find_suggestions
-        suggested == extensions
-      end
-
-      match_when_negated do
-        find_suggestions
-        suggested.none?
-      end
-
-      failure_message do
-        "expected to suggest extensions [#{extensions.join(', ')}], " \
-          "but got [#{suggested.join(', ')}]"
-      end
-
-      failure_message_when_negated do
-        "expected to not suggest extensions, but got [#{suggested.join(', ')}]"
-      end
-    end
-
-    let(:lockfile) do
-      create_file('Gemfile.lock', <<~LOCKFILE)
-        GEM
-          specs:
-            rake (13.0.1)
-            rspec (3.9.0)
-
-        PLATFORMS
-          ruby
-
-        DEPENDENCIES
-          rake (~> 13.0)
-          rspec (~> 3.7)
-      LOCKFILE
-    end
-
-    before do
-      create_file('example.rb', <<~RUBY)
-        # frozen_string_literal: true
-
-        puts 'ok'
-      RUBY
-
-      # Ensure that these specs works in CI, since the feature is generally
-      # disabled in when ENV['CI'] is set.
-      allow(ENV).to receive(:fetch).and_call_original
-      allow(ENV).to receive(:fetch).with('CI', nil).and_return(false)
-
-      # Mock the lockfile to be parsed by bundler
-      allow(Bundler).to receive(:default_lockfile)
-        .and_return(lockfile ? Pathname.new(lockfile) : nil)
-    end
-
-    context 'when bundler is not loaded' do
-      before { hide_const('Bundler') }
-
-      it 'does not show the suggestion' do
-        expect { cli.run(['example.rb']) }.not_to suggest_extensions
-        expect($stderr.string.blank?).to be(true)
-      end
-    end
-
-    context 'when there are gems to suggest' do
-      context 'that are not loaded' do
-        it 'shows the suggestion' do
-          expect { cli.run(['example.rb']) }.to suggest_extensions('rubocop-rake', 'rubocop-rspec')
-        end
-      end
-
-      context 'that are dependencies' do
-        let(:gemfile) do
-          create_file('Gemfile', <<~RUBY)
-            gem 'rspec'
-            gem 'rake'
-            gem 'rubocop-rspec'
-            gem 'rubocop-rake'
-          RUBY
-        end
-
-        before do
-          create_file('Gemfile.lock', <<~TEXT)
-            GEM
-              remote: https://rubygems.org/
-              specs:
-                rake (13.0.1)
-                rspec (3.9.0)
-                  rspec-core (~> 3.9.0)
-                rspec-core (3.9.3)
-                rubocop-rake (0.5.1)
-                rubocop-rspec (2.0.1)
-
-            DEPENDENCIES
-              rake (~> 13.0)
-              rspec (~> 3.7)
-              rubocop-rake (~> 0.5)
-              rubocop-rspec (~> 2.0.0)
-          TEXT
-        end
-
-        it 'does not show the suggestion' do
-          expect { cli.run(['example.rb']) }.not_to suggest_extensions
-        end
-      end
-
-      context 'that some are dependencies' do
-        let(:gemfile) do
-          create_file('Gemfile', <<~RUBY)
-            gem 'rspec'
-            gem 'rake'
-            gem 'rubocop-rake'
-          RUBY
-        end
-
-        before do
-          create_file('Gemfile.lock', <<~TEXT)
-            GEM
-              remote: https://rubygems.org/
-              specs:
-                rake (13.0.1)
-                rspec (3.9.0)
-                  rspec-core (~> 3.9.0)
-                rspec-core (3.9.3)
-                rubocop-rake (0.5.1)
-
-            DEPENDENCIES
-              rake (~> 13.0)
-              rspec (~> 3.7)
-              rubocop-rake (~> 0.5)
-          TEXT
-        end
-
-        it 'only suggests unused gems' do
-          expect { cli.run(['example.rb']) }.to suggest_extensions('rubocop-rspec')
-        end
-      end
-
-      context 'that are added by dependencies' do
-        let(:lockfile) do
-          create_file('Gemfile.lock', <<~TEXT)
-            GEM
-              specs:
-                rake (13.0.1)
-                rspec (3.9.0)
-                shared-gem (1.0.0)
-                  rubocop-rake (0.5.1)
-                  rubocop-rspec (2.0.1)
-
-            DEPENDENCIES
-              rake (~> 13.0)
-              rspec (~> 3.7)
-              shared-gem (~> 1.0.0)
-          TEXT
-        end
-
-        it 'does not show the suggestion' do
-          expect { cli.run(['example.rb']) }.not_to suggest_extensions
-        end
-      end
-    end
-
-    context 'when gems with suggestions are not primary dependencies' do
-      let(:lockfile) do
-        create_file('Gemfile.lock', <<~LOCKFILE)
-          GEM
-            specs:
-              shared-gem (1.0.0)
-                rake (13.0.1)
-                rspec (3.9.0)
-
-          PLATFORMS
-            ruby
-
-          DEPENDENCIES
-            shared-gem (~> 1.0)
-        LOCKFILE
-      end
-
-      it 'does not show the suggestion' do
-        expect { cli.run(['example.rb']) }.not_to suggest_extensions
-      end
-    end
-
-    context 'when there are multiple gems loaded that have the same suggestion' do
-      let(:lockfile) do
-        create_file('Gemfile.lock', <<~LOCKFILE)
-          GEM
-            specs:
-              rspec (3.9.0)
-              rspec-rails (4.0.1)
-
-          PLATFORMS
-            ruby
-
-          DEPENDENCIES
-            rspec (~> 3.9)
-            rspec-rails (~> 4.0)
-        LOCKFILE
-      end
-
-      it 'shows the suggestion' do
-        expect { cli.run(['example.rb']) }.to suggest_extensions('rubocop-rspec')
-      end
-    end
-
-    context 'with AllCops/SuggestExtensions: false' do
-      before do
-        create_file('.rubocop.yml', <<~YAML)
-          AllCops:
-            SuggestExtensions: false
-        YAML
-      end
-
-      it 'does not show the suggestion' do
-        expect { cli.run(['example.rb']) }.not_to suggest_extensions
-      end
-    end
-
-    context 'when an extension is disabled in AllCops/SuggestExtensions' do
-      before do
-        create_file('.rubocop.yml', <<~YAML)
-          AllCops:
-            SuggestExtensions:
-              rubocop-rake: false
-        YAML
-      end
-
-      it 'show the suggestion for non-disabled extensions' do
-        expect { cli.run(['example.rb']) }.to suggest_extensions('rubocop-rspec')
-      end
-    end
-
-    context 'when in CI mode' do
-      before { allow(ENV).to receive(:fetch).with('CI', nil).and_return(true) }
-
-      it 'does not show the suggestion' do
-        expect { cli.run(['example.rb']) }.not_to suggest_extensions
-      end
-    end
-
-    context 'when given --only' do
-      it 'does not show the suggestion' do
-        expect { cli.run(['example.rb', '--only', 'Style/Alias']) }.not_to suggest_extensions
-      end
-    end
-
-    context 'when given --debug' do
-      it 'does not show the suggestion' do
-        expect { cli.run(['example.rb', '--debug']) }.not_to suggest_extensions
-      end
-    end
-
-    context 'when given --list-target-files' do
-      it 'does not show the suggestion' do
-        expect { cli.run(['example.rb', '--list-target-files']) }.not_to suggest_extensions
-      end
-    end
-
-    context 'when given --out' do
-      it 'does not show the suggestion' do
-        expect { cli.run(['example.rb', '--out', 'output.txt']) }.not_to suggest_extensions
-      end
-    end
-
-    context 'when given --stdin' do
-      it 'does not show the suggestion' do
-        $stdin = StringIO.new('p $/')
-        expect { cli.run(['--stdin', 'example.rb']) }.not_to suggest_extensions
-      ensure
-        $stdin = STDIN
-      end
-    end
-
-    context 'when given a non-supported formatter' do
-      it 'does not show the suggestion' do
-        expect { cli.run(['example.rb', '--format', 'simple']) }.not_to suggest_extensions
-      end
-    end
-
-    context 'when given an invalid path' do
-      it 'does not show the suggestion' do
-        expect { cli.run(['example1.rb']) }.not_to suggest_extensions
-      end
-    end
-  end
-
   describe 'info severity' do
     let(:code) do
       <<~RUBY
